requirements.txt
# Note if using Conda it is recommended to install torch separately.
# For most of testing the following commands were run to set up the environment
# This was tested with torch==1.11.0
# conda create -n structure_recovery_gpu python=3.9
# conda activate structure_recovery_gpu
# conda install pytorch=1.11 cudatoolkit=11.1 -c pytorch -c nvidia
# pip install -r requirements.txt

# --------- pytorch --------- #
torch>=1.11.0
torchvision>=0.11.0
pytorch-lightning>=1.5.10
torchmetrics>=0.7.0
# pytorch-fast-transformers>=0.4.0

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0
hydra-submitit-launcher

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# tensorboard

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
pudb            # debugger

matplotlib
numpy
sdeint
torchdyn
scipy
pot
